+++ src/components/Home.js
@@ -1,0 +1,9 @@
+function Home() {
+  return (
+    <>
+    <h1>Home page</h1>
+    </>
+  )
+}
+
+export default Home;
